refactor: added confirmation window when delete and display No Todo Item if todoItems is empty

+++ src/components/TodoItem.js
@@ -10,7 +10,13 @@ const TodoItem = (props) => {
     };
 
     const deleteItem = () => {
-        dispatch(deleteTodoItem(props.index));
+        if(window.confirm("Are you sure you want to remove this item?")){
+            dispatch(deleteTodoItem(props.index));
+            window.alert("Todo Item removed successfully");
+        }
+        else{
+            window.alert("Todo item remove failed")
+        }
     };
 
     return (
